backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.8.5
supabase==2.15.3
